refactor: consolidate historical tests into domain-specific test files
Move guard tests from standalone historical fix files into their natural
locations: base test classes for integration tests and existing test
modules for unit tests. This provides regression coverage without test
duplication while keeping guards close to the code they protect.
