+++ tests/unit/test_plotting/test_quick_plot.py
@@ -500,6 +500,55 @@ class TestQuickPlot:
         with pytest.raises(TypeError, match=r"QuickPlot requires at least 1"):
             pybamm.QuickPlot([])
 
+    def test_extrapolation_time_clipping(self):
+        # Guard: time extrapolation should clip to valid range (PR #4991)
+        model = pybamm.lithium_ion.SPM()
+        sim = pybamm.Simulation(model)
+        sol = sim.solve([0, 100])
+
+        quick_plot = pybamm.QuickPlot(sol, ["Voltage [V]"])
+
+        quick_plot.plot(200)
+        assert quick_plot.fig is not None
+
+        quick_plot.plot(-50)
+        assert quick_plot.fig is not None
+
+        pybamm.close_plots()
+
+    def test_solution_starting_nonzero_time(self):
+        # Guard: solutions not starting at t=0 must work
+        model = pybamm.lithium_ion.SPM()
+        sim = pybamm.Simulation(model)
+
+        sol = sim.solve([50, 150])
+
+        quick_plot = pybamm.QuickPlot(sol, ["Voltage [V]"])
+        quick_plot.plot(100)
+
+        assert quick_plot.fig is not None
+        assert quick_plot.min_t_unscaled == 50
+        assert quick_plot.max_t_unscaled == 150
+
+        pybamm.close_plots()
+
+    def test_spatial_variable_interpolation_no_nan(self):
+        # Guard: spatial variables must interpolate without NaN (PR #4841)
+        model = pybamm.lithium_ion.SPMe()
+        sim = pybamm.Simulation(model)
+        sol = sim.solve([0, 100])
+
+        quick_plot = pybamm.QuickPlot(sol, ["Electrolyte concentration [mol.m-3]"])
+        quick_plot.plot(50)
+
+        plot_data = quick_plot.plots[("Electrolyte concentration [mol.m-3]",)][0][0]
+        ydata = plot_data.get_ydata()
+
+        assert not np.any(np.isnan(ydata))
+        assert not np.any(np.isinf(ydata))
+
+        pybamm.close_plots()
+
     def test_model_with_inputs(self):
         parameter_values = pybamm.ParameterValues("Chen2020")
         model = pybamm.lithium_ion.SPMe()
